Non-manifold error unreproducible.
Error appears to have been result of broken fold implementation and when
that is corrected it is not possible to reproduce that non-manifold
event (which is the expected behaviour with regularised set operations
on nef polyhedra).

When / if this issue is reproduced these tests will need to be expanded
to reliably reproduce the issue and then the possible fix is to prune
the non-manifold features when !is_simple.

This should be safe in all cases as these very small features would just
become swarf.