Add test case for duplicate point in path.
While writing another test case this fault was noticed. The faulty
comparison operators for point_3 and vector_3 allowed this fault to do
undetected.

The fault is not actually related to moving backwards, but the faulty
comparison operators themselves - which rely on normal double comparison
when such comparisons are inappropriate causing the check for an exact
end point to fail and a point that is duplicate to within acceptable
tolerances to be added.