Improve rotary path expansion.
Implemented expand_rotary and in doing so, implemented primitives to
measure path length for rotary axes in a pseudo cartesian system as
defined for pure rotary motion planning.

This then allowed the code for nearly-pure rotary motion (large angular
motion combined with a small linear move) to be improved. Now the path
is oversampled rather than undersampled in these circumstances.

This has the unfortunate side-effect of making the tests which exercise
this behaviour run unbearably long.

This is a sort term situation as the simulation code is scheduled to be
improved by removing the small per-step subtraction operations with a
combined tool path union operation and a second simulation step that
will operate over the entire path.