Export more of the interface
Preparing to implement rs274ngc js bindings + the ability for the js
components to expand path motion (some of the previous internal cxxcam
operations).

This is to allow both gcode generation with object updates for motion
planning AND gcode parsing with (optional) object updates for other
forms of analysis and rendering (e.g. WebGL backplot renderer).