Actually correct CW and CCW arcs.
Core issue identified, where delta_theta == 0, previous code failed to account for sign of delta theta in different arc
orientations, resulting in the hackish fix by alternating the sign of step_dt.

A more complete review at the time would have noticed that step_dt is a sign-adjusted rads_per_step, and that the core
issue was higher in the code.

Corrected to use correct sign of delta_theta when start and end points of arc are identical and delta_theta is assigned
to 2pi.